// frontend/lib/api/types.ts
export type Subject = {
  id: string;
  name: string;
  description: string | null;
  color: string | null;
  icon: string | null;
  created_at: string;
  archived_at: string | null;
  curriculum_subject_id: string | null;
};